backend/requirements.txt
Flask
Flask-Cors
pymongo
bcrypt
PyJWT
opencv-python
numpy
python-dotenv
